Resume_Analyser/Backend/requirements.txt
fastapi
uvicorn[standard]
spacy
sentence-transformers
scikit-learn
textstat
pdfminer.six
docx2txt
python-multipart
requests
